Add moving state to name selection
Previously multiple clicks on the arrows would cause the name selector to hit the end of the names elements. I set the name selector to jump at end of transition so multiple clicks were changing the index but not allowing the selector to jump when the edge names were reached.

The jumping is necessary to allow the name selector to be an infinite carousel.